Split generic type validation
